requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.5
aioresponses==0.7.7
aiosignal==1.3.1
attrs==23.2.0
awscli==1.36.5
backoff==2.2.1
beautifulsoup4==4.12.3
bing-image-downloader==1.1.2
boto3==1.35.64
botocore==1.35.64
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
docker==6.1.3
docutils==0.16
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
google_images_download==2.8.0
googlesearch-python==1.2.3
GPUtil==1.4.0
h11==0.14.0
idna==3.6
inference-cli==0.9.17
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.3
pendulum==3.0.0
pillow==10.2.0
prettytable==3.12.0
propcache==0.2.0
psutil==6.1.0
PuLP==2.9.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.31.0
rich==13.5.2
rpds-py==0.21.0
rsa==4.7.2
s3transfer==0.10.3
scipy==1.14.1
selenium==4.18.1
simple-image-download==0.2
six==1.16.0
skypilot==0.4.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
supervision==0.25.0
tabulate==0.9.0
time-machine==2.16.0
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
webdriver-manager==4.0.1
websocket-client==1.8.0
wheel==0.45.0
wsproto==1.2.0
yarl==1.17.2
pyyaml
ultralytics
scikit-learn
opencv-python
torch
numpy
wandb
tensorflow
tensorflowjs